new_rag_system/requirements.txt
fastapi
uvicorn[standard]
streamlit
langchain
sqlalchemy
psycopg2-binary
chromadb-client
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
PyYAML
sentence-transformers
pypdf
python-docx
tiktoken
openai
pytest
python-multipart
